LDA/app/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blis==1.0.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
cymem==2.0.11
exceptiongroup==1.3.0
fastapi==0.116.1
flatbuffers==25.2.10
fonttools==4.59.1
idna==3.10
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.3
murmurhash==1.0.13
numpy==2.2.6
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
preshed==3.0.10
protobuf==4.25.8
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
scipy==1.15.3
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
sounddevice==0.5.2
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.47.2
thinc==9.1.1
tqdm==4.67.1
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
